+++ src/main/java/net/spacerulerwill/skygrid_reloaded/ui/screen/CustomizeSkyGridScreen.java
@@ -5,6 +5,7 @@ import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.screen.world.CreateWorldScreen;
 import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.gui.widget.ClickableWidget;
+import net.minecraft.client.gui.widget.CyclingButtonWidget;
 import net.minecraft.client.gui.widget.DirectionalLayoutWidget;
 import net.minecraft.client.gui.widget.ThreePartsLayoutWidget;
 import net.minecraft.client.world.GeneratorOptionsHolder;
@@ -12,11 +13,12 @@ import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.registry.entry.RegistryEntryList;
 import net.minecraft.screen.ScreenTexts;
 import net.minecraft.text.Text;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.BiomeKeys;
-import net.minecraft.world.biome.source.FixedBiomeSource;
+import net.minecraft.world.biome.source.CheckerboardBiomeSource;
 import net.minecraft.world.dimension.DimensionOptions;
 import net.minecraft.world.dimension.DimensionOptionsRegistryHolder;
 import net.minecraft.world.dimension.DimensionType;
@@ -26,6 +28,7 @@ import net.minecraft.world.gen.chunk.ChunkGenerator;
 import net.minecraft.world.gen.chunk.ChunkGeneratorSettings;
 import net.spacerulerwill.skygrid_reloaded.SkyGridReloaded;
 import net.spacerulerwill.skygrid_reloaded.ui.widget.ClickableWidgetList;
+import net.spacerulerwill.skygrid_reloaded.ui.widget.IntSlider;
 import net.spacerulerwill.skygrid_reloaded.util.WorldPresetExtension;
 import net.spacerulerwill.skygrid_reloaded.worldgen.SkyGridChunkGenerator;
 import net.spacerulerwill.skygrid_reloaded.worldgen.SkyGridChunkGeneratorConfig;
@@ -37,8 +40,11 @@ import java.util.List;
 import java.util.Map;
 
 public class CustomizeSkyGridScreen extends Screen {
+    public static final List<RegistryKey<DimensionOptions>> DIMENSIONS = List.of(DimensionOptions.OVERWORLD, DimensionOptions.NETHER, DimensionOptions.END);
+
     private static final Text TITLE_TEXT = Text.translatable("createWorld.customize.skygrid.title");
     private static final Text BLOCKS_TEXT = Text.translatable("createWorld.customize.skygrid.blocks");
+    private static final Text BiOMES_TEXT = Text.translatable("createWorld.customize.skygrid.biomes");
     private static final Text SPAWNERS_TEXT = Text.translatable("createWorld.customize.skygrid.spawners");
     private static final Text LOOT_TEXT = Text.translatable("createWorld.customize.skygrid.loot");
     private static final Text PRESETS_TEXT = Text.translatable("createWorld.customize.skygrid.presets");
@@ -46,13 +52,40 @@ public class CustomizeSkyGridScreen extends Screen {
 
     private final ThreePartsLayoutWidget layout = new ThreePartsLayoutWidget(this);
     private final CreateWorldScreen parent;
+    private final IntSlider biomeScaleSlider;
     private ClickableWidgetList body;
+    private CyclingButtonWidget<RegistryKey<DimensionOptions>> dimensionsSelector;
     private SkyGridConfig currentConfig = new SkyGridConfig(SkyGridReloaded.DEFAULT_PRESET.config());
-
+    private RegistryKey<DimensionOptions> currentDimension = DimensionOptions.OVERWORLD;
 
     public CustomizeSkyGridScreen(CreateWorldScreen parent) {
         super(TITLE_TEXT);
         this.parent = parent;
+        this.biomeScaleSlider = new IntSlider(0, 0, 100, 20, Text.translatable("createWorld.customize.skygrid.biome_scale"), 0, 62, this.getCurrentConfig().checkerboardBiomeSource.scale, newValue -> {
+            SkyGridChunkGeneratorConfig currentConfig = this.getCurrentConfig();
+            currentConfig.checkerboardBiomeSource = new CheckerboardBiomeSource(
+                    RegistryEntryList.of(currentConfig.checkerboardBiomeSource.getBiomes().stream().toList()),
+                    newValue
+            );
+        });
+    }
+
+    private SkyGridChunkGeneratorConfig getCurrentConfig() {
+        SkyGridChunkGeneratorConfig config;
+        if (this.currentDimension == DimensionOptions.OVERWORLD) {
+            config = this.currentConfig.overworldConfig();
+        } else if (this.currentDimension == DimensionOptions.NETHER) {
+            config = this.currentConfig.netherConfig();
+        } else if (this.currentDimension == DimensionOptions.END) {
+            config = this.currentConfig.endConfig();
+        } else {
+            throw new IllegalStateException("Current dimension is not one of overworld, nether or end: " + this.currentDimension.getValue().toTranslationKey());
+        }
+        return config;
+    }
+
+    public void updateBiomeScaleSlider() {
+        this.biomeScaleSlider.setValue(this.getCurrentConfig().checkerboardBiomeSource.scale);
     }
 
     @Override
@@ -60,31 +93,47 @@ public class CustomizeSkyGridScreen extends Screen {
         // Header for title
         this.layout.addHeader(TITLE_TEXT, this.textRenderer);
         // Body
-        List<ClickableWidget> firstRow = List.of(
+        List<ClickableWidget> firstRow = List.of(new CyclingButtonWidget.Builder<RegistryKey<DimensionOptions>>(value -> Text.translatable(value.getValue().toTranslationKey()))
+                .values(DIMENSIONS)
+                .build(0, 0, 158, 20, Text.translatable("createWorld.customize.skygrid.dimension"), ((button, dimension) -> {
+                    this.currentDimension = dimension;
+                    this.updateBiomeScaleSlider();
+                }))
+        );
+        List<ClickableWidget> secondRow = List.of(
                 ButtonWidget.builder(BLOCKS_TEXT, (button) -> {
                     if (this.client != null) {
-                        this.client.setScreen(new CustomizeBlocksScreen(this, this.currentConfig));
+                        this.client.setScreen(new CustomizeBlocksScreen(this, this.currentDimension, this.currentConfig));
                     }
                 }).build(),
                 ButtonWidget.builder(SPAWNERS_TEXT, (button) -> {
                     if (this.client != null) {
-                        this.client.setScreen(new CustomizeSpawnerScreen(this, this.currentConfig));
+                        this.client.setScreen(new CustomizeSpawnerScreen(this, this.currentDimension, this.currentConfig));
                     }
                 }).build()
         );
-        List<ClickableWidget> secondRow = List.of(
+        List<ClickableWidget> thirdRow = List.of(
+                ButtonWidget.builder(BiOMES_TEXT, (button) -> {
+                    if (this.client != null) {
+                        Registry<Biome> biomeRegistry = this.parent.getWorldCreator().getGeneratorOptionsHolder().getCombinedRegistryManager().getOrThrow(RegistryKeys.BIOME);
+                        this.client.setScreen(new SelectBiomesScreen(this, biomeRegistry, this.currentDimension, this.currentConfig));
+                    }
+                }).build(),
+                this.biomeScaleSlider
+        );
+        List<ClickableWidget> fourthRow = List.of(
                 ButtonWidget.builder(LOOT_TEXT, (button) -> {
                     if (this.client != null) {
-                        this.client.setScreen(new CustomizeLootScreen(this, this.currentConfig));
+                        this.client.setScreen(new CustomizeLootScreen(this, this.currentDimension, this.currentConfig));
                     }
                 }).build(),
                 ButtonWidget.builder(PRESETS_TEXT, (button) -> {
                     if (this.client != null) {
-                        this.client.setScreen(new SkyGridPresetsScreen(this.client, this));
+                        this.client.setScreen(new SkyGridPresetsScreen(this.client, this, this.parent.getWorldCreator().getGeneratorOptionsHolder().getCombinedRegistryManager()));
                     }
                 }).build()
         );
-        List<List<ClickableWidget>> rows = List.of(firstRow, secondRow);
+        List<List<ClickableWidget>> rows = List.of(firstRow, secondRow, thirdRow, fourthRow);
         this.body = this.layout.addBody(new ClickableWidgetList(this.client, rows, this.width, this.layout.getContentHeight(), this.layout.getHeaderHeight()));
         // Footer
         DirectionalLayoutWidget footerRow = DirectionalLayoutWidget.horizontal().spacing(8);
@@ -131,29 +180,38 @@ public class CustomizeSkyGridScreen extends Screen {
         );
 
         return (dynamicRegistryManager, dimensionsRegistryHolder) -> {
-            Registry<Biome> biomeRegistry = dynamicRegistryManager.getOrThrow(RegistryKeys.BIOME);
             Registry<ChunkGeneratorSettings> chunkGeneratorSettingsRegistry = dynamicRegistryManager.getOrThrow(RegistryKeys.CHUNK_GENERATOR_SETTINGS);
             Map<RegistryKey<DimensionOptions>, DimensionOptions> updatedDimensions = new HashMap<>(dimensionsRegistryHolder.dimensions());
             dimensionOptionsToChunkGeneratorConfigMap.forEach((dimensionOptionsRegistryKey, config) -> {
-                boolean hasNonZeroBlock = config.blocks().values().stream().anyMatch(weight -> weight > 0);
+                boolean hasNonZeroBlock = config.blocks.values().stream().anyMatch(weight -> weight > 0);
                 if (hasNonZeroBlock) {
-                    RegistryKey<Biome> biomeRegistryKey = null;
+                    // Chunk generator settings
                     RegistryKey<ChunkGeneratorSettings> chunkGeneratorSettingsRegistryKey = null;
                     if (dimensionOptionsRegistryKey == DimensionOptions.OVERWORLD) {
-                        biomeRegistryKey = BiomeKeys.PLAINS;
                         chunkGeneratorSettingsRegistryKey = ChunkGeneratorSettings.OVERWORLD;
-                    }
-                    else if (dimensionOptionsRegistryKey == DimensionOptions.NETHER) {
-                        biomeRegistryKey = BiomeKeys.NETHER_WASTES;
+                    } else if (dimensionOptionsRegistryKey == DimensionOptions.NETHER) {
                         chunkGeneratorSettingsRegistryKey = ChunkGeneratorSettings.NETHER;
-                    }
-                    else if (dimensionOptionsRegistryKey == DimensionOptions.END) {
-                        biomeRegistryKey = BiomeKeys.THE_END;
+                    } else if (dimensionOptionsRegistryKey == DimensionOptions.END) {
                         chunkGeneratorSettingsRegistryKey = ChunkGeneratorSettings.END;
                     }
-                    RegistryEntry<Biome> biomeRegistryEntry = biomeRegistry.getEntry(biomeRegistry.get(biomeRegistryKey));
+
+                    // Biome source
+                    if (config.checkerboardBiomeSource.getBiomes().isEmpty()) {
+                        // Select single biome
+                        Registry<Biome> biomeRegistry = dynamicRegistryManager.getOrThrow(RegistryKeys.BIOME);
+                        RegistryKey<Biome> biomeKey = null;
+                        if (dimensionOptionsRegistryKey == DimensionOptions.OVERWORLD) {
+                            biomeKey = BiomeKeys.PLAINS;
+                        } else if (dimensionOptionsRegistryKey == DimensionOptions.NETHER) {
+                            biomeKey = BiomeKeys.NETHER_WASTES;
+                        } else if (dimensionOptionsRegistryKey == DimensionOptions.END) {
+                            biomeKey = BiomeKeys.THE_END;
+                        }
+                        RegistryEntry<Biome> biomeRegistryEntry = biomeRegistry.getEntry(biomeRegistry.get(biomeKey));
+                        config.checkerboardBiomeSource = new CheckerboardBiomeSource(RegistryEntryList.of(biomeRegistryEntry), 1);
+                    }
                     RegistryEntry<ChunkGeneratorSettings> chunkGeneratorSettingsRegistryEntry = chunkGeneratorSettingsRegistry.getEntry(chunkGeneratorSettingsRegistry.get(chunkGeneratorSettingsRegistryKey));
-                    ChunkGenerator chunkGenerator = new SkyGridChunkGenerator(new FixedBiomeSource(biomeRegistryEntry), chunkGeneratorSettingsRegistryEntry, config);
+                    ChunkGenerator chunkGenerator = new SkyGridChunkGenerator(chunkGeneratorSettingsRegistryEntry, config);
                     DimensionOptions dimensionOptions = parent.getWorldCreator().getGeneratorOptionsHolder().selectedDimensions().dimensions().get(dimensionOptionsRegistryKey);
 
                     RegistryEntry<DimensionType> dimensionTypeRegistryEntry = dimensionOptions.dimensionTypeEntry();